Extract z.ai response parser, add tests, and changelog
Move parseZaiApiResponse into parser.ts so it loads without @raycast/api,
then cover the daily/weekly window disambiguation with unit tests. Add the
changelog entry for the z.ai weekly limits feature.
